+++ CheersToChinchillia/src/game/models/SlimChillia.java
@@ -1,10 +1,12 @@
 package game.models;
 
 public class SlimChillia extends Player {
+	//Attributes
 	private double hungerMultiplier = -.08;
 	private double happinessMultiplier = .3;
 	
-	public SlimChillia(String name, int health, int age, int happiness, int hunger, int attackDamage) {
+	// Constructor
+	public SlimChillia(String name, double health, int age, double happiness, double hunger, double attackDamage) {
 		this.setAge(age);
 		this.setName(name);
 		this.setHealth(health);
@@ -12,6 +14,7 @@ public class SlimChillia extends Player {
 		this.setAttackDamage(attackDamage);
 	}
 	
+	//Behaviors
 	public void performMultiplers() {
 		this.setHappiness(this.getHappiness()*happinessMultiplier);
 		this.setHunger(this.getHunger()*hungerMultiplier);
